+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/ControlledTag.test.jsx
@@ -1,0 +1,41 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import Select from 'terra-form-select/lib/Select';
+import styles from './common/Select.test.module.scss';
+
+const cx = classNames.bind(styles);
+
+class ControlledTag extends React.Component {
+  constructor() {
+    super();
+
+    this.state = { value: [] };
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  handleChange(value) {
+    this.setState({ value });
+  }
+
+  render() {
+    return (
+      <div className={cx('content-wrapper')}>
+        <Select
+          onChange={this.handleChange}
+          placeholder="Select a color"
+          variant="tag"
+          id="tag"
+          value={this.state.value}
+        >
+          <Select.Option value="blue" display="Blue" />
+          <Select.Option value="green" display="Green" />
+          <Select.Option value="purple" display="Purple" />
+          <Select.Option value="red" display="Red" />
+          <Select.Option value="violet" display="Violet" />
+        </Select>
+      </div>
+    );
+  }
+}
+
+export default ControlledTag;