+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/TagSelectPlaceholder.test.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import Select from 'terra-form-select/lib/Select';
+import styles from './common/Select.test.module.scss';
+
+const cx = classNames.bind(styles);
+
+const TagSelectPlaceholder = () => (
+  <div className={cx('content-wrapper')}>
+    <Select placeholder="Select a color" variant="tag" id="tag">
+      <Select.Option value="blue" display="Blue" />
+      <Select.Option value="green" display="Green" />
+      <Select.Option value="purple" display="Purple" />
+      <Select.Option value="red" display="Red" />
+      <Select.Option value="violet" display="Violet" />
+    </Select>
+    <button type="button" id="focusable">Focusable</button>
+  </div>
+);
+
+export default TagSelectPlaceholder;