+++ packages/terra-form-select/tests/wdio/native-select-field-spec.js
@@ -1,18 +1,18 @@
 Terra.describeViewports('Native Select Field', ['tiny'], () => {
   it('should display default Field', () => {
-    browser.url('/raw/tests/terra-form-select/form-select/native-select/default-field');
+    browser.url('/raw/tests/cerner-terra-core-docs/form-select/native-select/default-field');
 
     Terra.validates.element('default');
   });
 
   it('should display incomplete Field', () => {
-    browser.url('/raw/tests/terra-form-select/form-select/native-select/incomplete-field');
+    browser.url('/raw/tests/cerner-terra-core-docs/form-select/native-select/incomplete-field');
 
     Terra.validates.element('incomplete');
   });
 
   it('should display required Field', () => {
-    browser.url('/raw/tests/terra-form-select/form-select/native-select/required-field');
+    browser.url('/raw/tests/cerner-terra-core-docs/form-select/native-select/required-field');
 
     Terra.validates.element('required');
   });