+++ packages/terra-core-docs/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Migrate docs from terra-form-field.
   * Migrate docs from terra-divider.
   * Migrate docs from terra-form-fieldset.
+  * Migrate docs from terra-form-select.
   * Migrate docs from terra-form-radio.
   * Migrate docs from terra-form-input.
   * Migrate docs from terra-form-textarea.
@@ -36,7 +37,7 @@
   * Updated Arrange tests to be more reliable.
 
 * Added
-  * Migrate doc from terra-avatar
+  * Migrate docs from terra-avatar
 
 ## 1.9.0 - (September 21, 2021)