+++ packages/terra-form-select/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Help text div no longer has a tabIndex of -1.
   * Updated dependencies to not be hard-coded.
+  * Migrate doc to terra-core-docs.
 
 ## 6.35.0 - (September 21, 2021)